refactor(core): Phase 3 — extract FlowOrchestratorEngine; shrink HangfireFlowOrchestrator to a shim
Moves the entire flow execution lifecycle (trigger, step run, graph continuation, polling,
cancellation, timeout, retry) out of HangfireFlowOrchestrator and into the new
FlowOrchestratorEngine in FlowOrchestrator.Core. Hangfire's only job is now to extract
the job ID from PerformContext and delegate to the engine.

- IFlowOrchestrator: new Core interface (TriggerAsync, TriggerByScheduleAsync, RunStepAsync, RetryStepAsync)
- FlowOrchestratorEngine: runtime-neutral implementation; zero Hangfire references
- HangfireFlowOrchestrator: shrunk to a 35-line adapter shim
- FlowOrchestratorTelemetry: moved from Hangfire to Core (no Hangfire deps)
- IExecutionContext.JobId: added so runtime adapters can propagate job/message IDs without Hangfire types
- ITriggerContext.JobId: removed (superseded by IExecutionContext.JobId)
- StepInstance, TypedStepInstanceAdapter, ExecutionContext: implement new JobId property
- FlowOrchestrator.Core.csproj: added Microsoft.Extensions.Logging.Abstractions
