+++ src/FlowOrchestrator.Core/Execution/ITriggerContext.cs
@@ -7,6 +7,10 @@ namespace FlowOrchestrator.Core.Execution;
 /// that started the run. Passed to <see cref="IFlowExecutor.TriggerFlow"/> and
 /// <see cref="IFlowGraphPlanner.CreateEntrySteps"/> to bootstrap execution.
 /// </summary>
+/// <remarks>
+/// The runtime-specific job or message ID for the trigger is carried by <see cref="IExecutionContext.JobId"/>,
+/// which is set by the runtime adapter (e.g. the Hangfire shim) before invoking the engine.
+/// </remarks>
 public interface ITriggerContext : IExecutionContext
 {
     /// <summary>The flow definition being triggered.</summary>
@@ -14,10 +18,4 @@ public interface ITriggerContext : IExecutionContext
 
     /// <summary>The trigger event that fired (key, type, payload, headers).</summary>
     ITrigger Trigger { get; set; }
-
-    /// <summary>
-    /// The Hangfire background job ID of the trigger job, if available.
-    /// Stored on the run record for correlation with the Hangfire dashboard.
-    /// </summary>
-    string? JobId { get; set; }
 }
