+++ src/FlowOrchestrator.Hangfire/FlowOrchestratorInstrumentationExtensions.cs
@@ -1,3 +1,4 @@
+using FlowOrchestrator.Core.Execution;
 using OpenTelemetry.Metrics;
 using OpenTelemetry.Trace;
