+++ tests/FlowOrchestrator.Core.Tests/Execution/PollableStepHandlerTests.cs
@@ -186,6 +186,7 @@ public class PollableStepHandlerTests
         public string? PrincipalId { get; set; }
         public object? TriggerData { get; set; }
         public IReadOnlyDictionary<string, string>? TriggerHeaders { get; set; }
+        public string? JobId { get; set; }
         public DateTimeOffset ScheduledTime { get; set; }
         public string Type { get; set; }
         public string Key { get; }
